requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alembic==0.8.10           # via flask-migrate
babel==2.4.0              # via flask-babelex
blinker==1.4              # via flask-mail, flask-principal
click==6.7                # via flask
flask-admin==1.5.0
flask-babelex==0.9.3      # via flask-security
flask-login==0.4.0        # via flask-security
flask-mail==0.9.1         # via flask-security
flask-migrate==2.0.3
flask-principal==0.4.0    # via flask-security
flask-script==2.0.5       # via flask-migrate
flask-security==3.0.0
flask-sqlalchemy==2.1
flask-sslify==0.1.5
flask-wtf==0.14.2         # via flask-security
flask==0.12
itsdangerous==0.24        # via flask, flask-security
jinja2==2.9.5             # via flask, flask-babelex
mako==1.0.6               # via alembic
markupsafe==0.23          # via jinja2, mako
passlib==1.7.1            # via flask-security
pillow==5.1.0
pwgen==0.8.2.post0
python-editor==1.0.3      # via alembic
pytz==2017.2              # via babel
speaklater==1.3           # via flask-babelex
sqlalchemy==1.1.5         # via alembic, flask-sqlalchemy
werkzeug==0.11.15         # via flask
wtforms==2.1              # via flask-admin, flask-wtf
